DROOLS-7020 Update jbang catalog deps to fixed ver
Since Drools published version 8.x versions on Maven Central,
they are assumed to be the most up-to-date RELEASE
(regardless of the .Beta postfix).

As a consequence, tests are failing on non-cached resolutions.

At this point, is better to just fix an arbitrary 7.x .Final version
for the command line utilities.